S3/dogsClassifier/requirement.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
huggingface-hub==0.25.1
idna==3.10
Jinja2==3.1.4
lightning==2.4.0
lightning-utilities==0.11.7
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
numpy==2.0.2
packaging==24.1
pillow==10.4.0
protobuf==5.28.2
pytorch-lightning==2.4.0
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.5
sympy==1.13.3
tensorboardX==2.6.2.2
timm==1.0.9
torch==2.4.1
torchmetrics==1.4.2
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.13.0
